==> site/40_workload.qmd <==
---
title: Workloads 
---

These are experimental presentations of the data.

```{python}
#| echo: false
#| eval: true
import os
import sys
import json
import pandas as pd
import panel as pn
from bokeh.models.widgets.tables import NumberFormatter

pn.extension('tabulator')

# Load dataframe 
stacked_df = pd.read_csv('stacked_df.csv')

```

## Explore tabular data

```{python}
cols = ['instructor_name','ins_dept','TERM','COURSE','TITLE','wrkld_sample_type','sum_term_crse_wrkld_sample','wrkld_sample_note']
terms = [202610,202520]

# Strip whitespace from column names
stacked_df.columns = stacked_df.columns.str.strip()

df = stacked_df[stacked_df["TERM"].isin(terms)][cols]

bokeh_formatters = {
    'TERM': NumberFormatter(format='000000')
}

table = pn.widgets.Tabulator(df,
    show_index=False,
    formatters=bokeh_formatters
)

table
```
